DateTime support for tick-precision and/or Fixed64 encoding; fixed Int64 encoding bug.
git-svn-id: http://protobuf-net.googlecode.com/svn/trunk@157 b72047a4-3652-0410-9617-b3994939e97b

+++ protobuf-net/SerializationContext.Read.cs
@@ -279,20 +279,20 @@ namespace ProtoBuf
         public long DecodeInt64Fixed()
         {
             Fill(8, true);
-            int lo =
-                   ((int)ioBuffer[ioBufferIndex++])
-                | (((int)ioBuffer[ioBufferIndex++]) << 8)
-                | (((int)ioBuffer[ioBufferIndex++]) << 16)
-                | (((int)ioBuffer[ioBufferIndex++]) << 24),
+            uint lo =
+                   ((uint)ioBuffer[ioBufferIndex++])
+                | (((uint)ioBuffer[ioBufferIndex++]) << 8)
+                | (((uint)ioBuffer[ioBufferIndex++]) << 16)
+                | (((uint)ioBuffer[ioBufferIndex++]) << 24),
                 hi =
-                   ((int)ioBuffer[ioBufferIndex++])
-                | (((int)ioBuffer[ioBufferIndex++]) << 8)
-                | (((int)ioBuffer[ioBufferIndex++]) << 16)
-                | (((int)ioBuffer[ioBufferIndex++]) << 24);
+                   ((uint)ioBuffer[ioBufferIndex++])
+                | (((uint)ioBuffer[ioBufferIndex++]) << 8)
+                | (((uint)ioBuffer[ioBufferIndex++]) << 16)
+                | (((uint)ioBuffer[ioBufferIndex++]) << 24);
 
             position += 8;
-            long loL = (long)lo, hiL = (long)hi;
-            return (hiL << 32) | loL;
+            ulong loL = (ulong)lo, hiL = (ulong)hi;
+            return (long)((hiL << 32) | loL);
         }
         public int Read(int count)
         {
